+++ saleor/graphql/account/mutations/staff.py
@@ -533,9 +533,8 @@ class UserAvatarUpdate(BaseMutation):
     def perform_mutation(cls, _root, info, image):
         user = info.context.user
         image_data = info.context.FILES.get(image)
-        validate_image_file(image_data, "image")
+        validate_image_file(image_data, "image", AccountErrorCode)
         add_hash_to_file_name(image_data)
-
         if user.avatar:
             user.avatar.delete_sized_images()
             user.avatar.delete()